refactor: extract reusable categories section
